# chore(client): don't await tx until it was submitted 