Re #6530

Most of the original code was superseded/incorporated by #6538

The last optimization left is not starting await until first successful `submit`. This cuts number of requests in parallel at the beginning by half, which might be meaningful if the client starts with multiple pending txes to submit after period of being offline in a low bandwidth env.

I was a bit worried about extra latency, but it looks like it's OK - [the consensus has latency as well](https://github.com/fedimint/fedimint/discussions/6521), so client has more than enough time to make a another request. 

old:

```
### LATENCY FM PAY: min: 1.0s, avg: 1.1s, median: 1.1s, p90: 1.4s, max: 1.4s, sum: 11.2s
### LATENCY FM PAY: min: 1.0s, avg: 1.6s, median: 1.1s, p90: 6.1s, max: 6.1s, sum: 16.2s
```


new:

```
### LATENCY FM PAY: min: 1.0s, avg: 1.1s, median: 1.1s, p90: 1.2s, max: 1.2s, sum: 11.0s
### LATENCY FM PAY: min: 1.0s, avg: 1.1s, median: 1.1s, p90: 1.2s, max: 1.2s, sum: 11.2s
### LATENCY FM PAY: min: 1.0s, avg: 1.1s, median: 1.1s, p90: 1.7s, max: 1.7s, sum: 11.4s
```

<!--

# Code Review Policy

* CI must pass (enforced)
* 1 review is mandatory (enforced), 2 or more ideal
* If you believe your change is simple, and non-controversial enough, and you want
  to avoid merge conflicts, or blocking work before it gets enough reviews, label it with
  `needs further review` label and Merge it.

See https://github.com/fedimint/fedimint/blob/master/CONTRIBUTING.md#code-review-policy for
full description.

-->