@dpc 2024-12-10
@elsirion Seems like a likely cause of executor "getting stuck" on bad internet connections?