@dpc 2024-12-11
Backcompat tests failing with:

```
00:00:20     2024-12-10T23:39:18.451651Z DEBUG fedimint_cli: Command failed err=Reissue failed: Transaction not accepted "The transaction had an invalid input: 1-The note is already spent"
```

for fedimintd 0.2.1.

I vaguely remember that we had an API race condition like that in the past, and possibly the timing changes in this PR are triggering it.