@joschisan 2024-12-11
> I was a bit worried about not starting await before first succesful submit, but it looks like it's OK 

Why make this optimization? Why not already start the subscription?