@dpc 2024-12-11
> Why make this optimization? Why not already start the subscription?

It cuts the number of calls running in parallel by two, at least initially.